requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-crontab==0.7.1
channels==4.0.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cache & Queue
redis==5.0.1
celery==5.3.4
django-redis==5.4.0

# API Documentation
drf-spectacular==0.27.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.1
PyJWT==2.8.0

# Google AI
google-generativeai==0.3.2

# HTTP & API
requests==2.31.0
httpx==0.26.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
biopython==1.83

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
pylint==3.0.3

# Debugging
django-debug-toolbar==4.2.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk==1.39.1

# Utilities
python-slugify==8.0.1
Pillow==10.2.0
